requirements.txt
streamlit>=1.44.0
openai>=1.68.0
python-dotenv>=1.0.1
google-api-python-client>=2.164.0
google-auth>=2.38.0
reportlab>=4.3.1
